feat: implement timeout handling for blocked clients and refactor related classes

+++ src/main/java/record/BlockedClient.java
@@ -2,14 +2,14 @@ package record;
 
 import java.nio.channels.SocketChannel;
 
-public record BlockedClient(SocketChannel client, long timeoutEndMillis) {
-    private static final long NO_EXPIRY = 0L;
+public record BlockedClient(SocketChannel client, double timeoutEndMillis) {
+    private static final double NO_EXPIRY = 0L;
 
     public boolean isExpired() {
         return timeoutEndMillis != 0 && System.currentTimeMillis() >= timeoutEndMillis;
     }
 
-    public static BlockedClient withExpiry(SocketChannel client, long timeoutEndMillis) {
+    public static BlockedClient withExpiry(SocketChannel client, double timeoutEndMillis) {
         return new BlockedClient(client, timeoutEndMillis);
     }
